esp/home/requirements.txt
# Python dependencies for CommuniFarm ESPHome setup
PyYAML>=6.0
